fig2pptx/requirements.txt
wheel>=0.22
pytest
numpy
matplotlib
python-pptx
jupyter